validate in middleware instead
