+++ evap/evaluation/tests/tools.py
@@ -116,77 +116,8 @@ class SimpleTestCase(SeedBakerMixin, ResetLanguageOnTearDownMixin, django.test.S
     pass
 
 
-class InvalidHtmlError(Exception):
-    def __init__(self, path: str, html: str, errors: list) -> None:
-        self.path = path
-        self.html = html
-        self.errors = errors
-
-    def format_error(self, error) -> str:
-        s = f"error: {error['message']}\n"
-        if "extract" in error:
-            s += f"context:\n{error['extract']}"
-        return s
-
-    def __str__(self) -> str:
-        errors = "\n".join(self.format_error(e) for e in self.errors)
-        return f"html validation failed for {self.path}:\n{errors}"
-
-
-class ValidatingTestApp(django_webtest.DjangoTestApp):
-    IGNORED_ERROR_PATTERNS = (
-        "autocomplete",
-        "not allowed as child of element “span”",
-        # accessibility
-        "aria",
-        'same "nearest ancestor autofocus scoping root element"',
-        "Every active “role=tab” element must have a corresponding “role=tabpanel” element",
-        "An element with “role=tab” must be contained in, or owned by, an element with the “role” value “tablist”",
-        # custom attributes
-        "custom-success",
-        "reload-on-success",
-        "Attribute “tomselect-no-sort” not allowed on element “select”",
-        # table correctness
-        "seen in “table”",
-        "not allowed as child of element “tr”",
-        # column count
-        "999 columns",
-        "999 established",
-    )
-
-    vnu_url: str | None = settings.VNU_URL
-    if vnu_url is None:
-        warnings.warn("Cannot validate HTML, VNU_URL not provided in settings", RuntimeWarning, stacklevel=2)
-
-    def do_request(self, req, *args, **kwargs) -> django_webtest.DjangoWebtestResponse:
-        response = super().do_request(req, *args, **kwargs)
-        assert isinstance(response, django_webtest.DjangoWebtestResponse)
-        if response.content_type == "text/html" and response.text:
-            self.validate_html(req.path, response.text)
-        return response
-
-    def validate_html(self, path: str, html: str) -> None:
-        if self.vnu_url is None:
-            return
-
-        errors = requests.post(
-            self.vnu_url,
-            params={
-                "out": "json",
-                "level": "error",
-                "filterpattern": "|".join(f".*{p}.*" for p in self.IGNORED_ERROR_PATTERNS),
-            },
-            headers={"Content-Type": "text/html"},
-            data=html,
-            timeout=10,
-        ).json()["messages"]
-
-        if errors:
-            raise InvalidHtmlError(path, html, errors)
-
-
 class WebTest(SeedBakerMixin, ResetLanguageOnTearDownMixin, django_webtest.WebTest):
-    app_class = ValidatingTestApp
+    pass
 
 
 def to_querydict(dictionary):
